madcat@192.168.1.123/requirements.txt
python-dotenv
json
numpy
requests